refactor step 3
